Support a persistent connection with the server
By adding support for a persistent connection, two features are added:
* Multiple commands can be sent on the same connection. This reduces
  the amount of time a commend takes before it takes effect (relatively
  small amount but can be percieved when running on low-power devices).
* Allows the client to get a stream of the height and speed changes.

In order to better support non-idasen-controller clients, the logic to
determine the height to set was moved from startup, to the set_height
command generation. This way, the client does not need to determine the
raw height value themselves.

Note that this implementation may be slightly buggy with respects to
threading/async tasks. The author is not generally familiar with Python
and definitely does not know it's threading model or the way asyncio
is intended to be used.